[svn] Added GPL license and got rid of 2 minor warnings.

+++ src/org/codejive/gui4gl/fonts/FixedPitchTextureFont.java
@@ -1,6 +1,23 @@
 /*
+ * [gui4gl] OpenGL game-oriented GUI library
+ * 
+ * Copyright (C) 2003 Steven Lagerweij
+ * 
+ * This library is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as published by the
+ * Free Software Foundation; either version 2.1 of the License, or (at your
+ * option) any later version.
+ * 
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License
+ * for more details.
+ * 
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this library; if not, write to the Free Software Foundation,
+ * Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
+ * 
  * Created on Dec 10, 2003
- *
  */
 package org.codejive.gui4gl.fonts;
 
@@ -10,7 +27,7 @@ import org.codejive.utils4gl.RenderContext;
 
 /**
  * @author steven
- * @version $Revision: 194 $
+ * @version $Revision: 196 $
  */
 public class FixedPitchTextureFont implements Font {
 	
@@ -79,8 +96,8 @@ public class FixedPitchTextureFont implements Font {
 	 */
 	public FixedPitchTextureFont(String _sCharacterSet, int _lCharactersPerRow, int _lTextureId, int _lTextureWidth, int _lTextureHeight, float _fXOffset, float _fYOffset, float _fCharacterWidth, float _fCharacterHeight, float _fBaseLine, boolean _bHasLowercase) {
 		m_lTextureId = _lTextureId;
-		m_fTexturePixelWidth = 1.0f / (float)_lTextureWidth;
-		m_fTexturePixelHeight = 1.0f / (float)_lTextureHeight;
+		m_fTexturePixelWidth = 1.0f / _lTextureWidth;
+		m_fTexturePixelHeight = 1.0f / _lTextureHeight;
 		
 		m_fCharWidth = _fCharacterWidth;
 		m_fCharHeight = _fCharacterHeight;
@@ -199,6 +216,9 @@ public class FixedPitchTextureFont implements Font {
 
 /*
  * $Log$
+ * Revision 1.2  2003/12/11 13:59:19  tako
+ * Added GPL license and got rid of 2 minor warnings.
+ *
  * Revision 1.1  2003/12/11 12:57:07  steven
  * Allows for texture mapped (transparant) fonts.
  * You must load the texture yourself and you must use
